Rename Test files to follow normal pattern
