+++ src/SFA.DAS.FAT.Web.UnitTests/Models/CoursesViewModelTests/WhenGettingOrderByName.cs
@@ -1,40 +1,41 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using AutoFixture.NUnit3;
+﻿using AutoFixture.NUnit3;
 using FluentAssertions;
 using NUnit.Framework;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
 
 namespace SFA.DAS.FAT.Web.UnitTests.Models.CoursesViewModelTests
 {
-    class WhenOrderingLinksByRelevance
+    class WhenGettingOrderByName
     {
         [Test]
-        public void Then_Adds_OrderBy_Relevance_To_Query_String()
+        public void Then_Adds_OrderBy_Name_To_Query_String()
         {
             // Arrange Act
-            var model = CoursesViewModelFactory.BuildModel(new List<Guid>(), null, new List<int>(), "relevance");
+            var model = CoursesViewModelFactory.BuildModel(new List<Guid>(), null, new List<int>(), "name");
             // Assert
-            model.OrderByRelevance.Should().Be("?orderby=relevance");
+            model.OrderByName.Should().Be("?orderby=name");
         }
 
         [Test, AutoData]
-        public void Then_Adds_Keyword_To_Query_String(string keyword)
+        public void And_Then_Adds_Keyword_To_Query_String(string keyword)
         {
             // Arrange Act
-            var model = CoursesViewModelFactory.BuildModel(new List<Guid>(), keyword, new List<int>(), "relevance");
+            var model = CoursesViewModelFactory.BuildModel(new List<Guid>(), keyword, new List<int>(), "name");
             // Assert
-            model.OrderByRelevance.Should​().Be($"?keyword={model.Keyword}&orderby=relevance");
+            model.OrderByName.Should​().Be($"?keyword={model.Keyword}&orderby=name");
         }
-
+        
         [Test, AutoData]
         public void And_Then_Adds_Sectors_To_Query_String(List<Guid> selectedRoutes)
         {
             // Arrange 
             var buildSectorLink = "";
-
+            
             // Act
-            var model = CoursesViewModelFactory.BuildModel(selectedRoutes, null, new List<int>(), "relevance");
+            var model = CoursesViewModelFactory.BuildModel(selectedRoutes, null, new List<int>(), "name");
             foreach (var selectedRoute in selectedRoutes)
             {
                 var separator = string.IsNullOrEmpty(buildSectorLink) ? "?" : "&";
@@ -42,9 +43,8 @@ namespace SFA.DAS.FAT.Web.UnitTests.Models.CoursesViewModelTests
             }
 
             // Assert​
-            model.OrderByRelevance.Should().Be($"{buildSectorLink}&orderby=relevance");
+            model.OrderByName.Should().Be($"{buildSectorLink}&orderby=name");
         }
-
         [Test, AutoData]
         public void And_Then_Adds_Levels_To_Query_String(List<int> selectedLevels)
         {
@@ -52,8 +52,8 @@ namespace SFA.DAS.FAT.Web.UnitTests.Models.CoursesViewModelTests
             var buildLevelsLink = "";
 
             // Act
-            var model = CoursesViewModelFactory.BuildModel(new List<Guid>(), null, selectedLevels, "relevance");
-
+            var model = CoursesViewModelFactory.BuildModel(new List<Guid>(), null, selectedLevels, "name");
+            
             foreach (var selectedLevel in selectedLevels)
             {
                 var separator = string.IsNullOrEmpty(buildLevelsLink) ? "?" : "&";
@@ -61,23 +61,17 @@ namespace SFA.DAS.FAT.Web.UnitTests.Models.CoursesViewModelTests
             }
 
             // Assert​
-            model.OrderByRelevance.Should().Be($"{buildLevelsLink}&orderby=relevance");
+            model.OrderByName.Should().Be($"{buildLevelsLink}&orderby=name");
         }
-
         [Test, AutoData]
-        public void Then_Adds_Sectors_And_Levels_To_Query_String(List<int> selectedLevels, List<Guid> selectedRoutes)
+        public void And_Then_Adds_Sectors_And_Levels_To_Query_String(List<int> selectedLevels, List<Guid> selectedRoutes)
         {
             // Arrange 
             var buildLevelsLink = "";
             var buildSectorLink = "";
-
+           
             // Act
-            var model = CoursesViewModelFactory.BuildModel(selectedRoutes, null, selectedLevels, "relevance");
-            foreach (var selectedLevel in selectedLevels)
-            {
-                var separator = string.IsNullOrEmpty(buildLevelsLink) ? "?" : "&";
-                buildLevelsLink += model.SelectedLevels != null && model.SelectedLevels.Any() ? $"{separator}levels=" + string.Join("levels=", selectedLevel) : "";
-            }
+            var model = CoursesViewModelFactory.BuildModel(selectedRoutes, null, selectedLevels, "name");
 
             foreach (var selectedRoute in selectedRoutes)
             {
@@ -85,8 +79,14 @@ namespace SFA.DAS.FAT.Web.UnitTests.Models.CoursesViewModelTests
                 buildSectorLink += model.SelectedSectors != null && model.SelectedSectors.Any() ? $"{separator}sectors=" + string.Join("&sectors=", selectedRoute) : "";
             }
 
+            foreach (var selectedLevel in selectedLevels)
+            {
+                var separator = string.IsNullOrEmpty(buildLevelsLink) ? "?" : "&";
+                buildLevelsLink += model.SelectedLevels != null && model.SelectedLevels.Any() ? $"{separator}levels=" + string.Join("&levels=", selectedLevel) : "";
+            }
+
             // Assert
-            model.OrderByRelevance.Should().Be($"{buildSectorLink}{buildLevelsLink}&orderby=relevance");
+            model.OrderByName.Should().Be($"{buildSectorLink}{buildLevelsLink}&orderby=name");
         }
     }
 }
