+++ src/SFA.DAS.FAT.Web.UnitTests/Models/CoursesViewModelTests/WhenGettingOrderByRelevance.cs
@@ -1,41 +1,40 @@
-﻿using AutoFixture.NUnit3;
-using FluentAssertions;
-using NUnit.Framework;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
+using AutoFixture.NUnit3;
+using FluentAssertions;
+using NUnit.Framework;
 
 namespace SFA.DAS.FAT.Web.UnitTests.Models.CoursesViewModelTests
 {
-    class WhenOrderingLinksByName
+    class WhenGettingOrderByRelevance
     {
         [Test]
-        public void Then_Adds_OrderBy_Name_To_Query_String()
+        public void Then_Adds_OrderBy_Relevance_To_Query_String()
         {
             // Arrange Act
-            var model = CoursesViewModelFactory.BuildModel(new List<Guid>(), null, new List<int>(), "name");
+            var model = CoursesViewModelFactory.BuildModel(new List<Guid>(), null, new List<int>(), "relevance");
             // Assert
-            model.OrderByName.Should().Be("?orderby=name");
+            model.OrderByRelevance.Should().Be("?orderby=relevance");
         }
 
         [Test, AutoData]
-        public void And_Then_Adds_Keyword_To_Query_String(string keyword)
+        public void Then_Adds_Keyword_To_Query_String(string keyword)
         {
             // Arrange Act
-            var model = CoursesViewModelFactory.BuildModel(new List<Guid>(), keyword, new List<int>(), "name");
+            var model = CoursesViewModelFactory.BuildModel(new List<Guid>(), keyword, new List<int>(), "relevance");
             // Assert
-            model.OrderByName.Should​().Be($"?keyword={model.Keyword}&orderby=name");
+            model.OrderByRelevance.Should​().Be($"?keyword={model.Keyword}&orderby=relevance");
         }
-        
+
         [Test, AutoData]
         public void And_Then_Adds_Sectors_To_Query_String(List<Guid> selectedRoutes)
         {
             // Arrange 
             var buildSectorLink = "";
-            
+
             // Act
-            var model = CoursesViewModelFactory.BuildModel(selectedRoutes, null, new List<int>(), "name");
+            var model = CoursesViewModelFactory.BuildModel(selectedRoutes, null, new List<int>(), "relevance");
             foreach (var selectedRoute in selectedRoutes)
             {
                 var separator = string.IsNullOrEmpty(buildSectorLink) ? "?" : "&";
@@ -43,8 +42,9 @@ namespace SFA.DAS.FAT.Web.UnitTests.Models.CoursesViewModelTests
             }
 
             // Assert​
-            model.OrderByName.Should().Be($"{buildSectorLink}&orderby=name");
+            model.OrderByRelevance.Should().Be($"{buildSectorLink}&orderby=relevance");
         }
+
         [Test, AutoData]
         public void And_Then_Adds_Levels_To_Query_String(List<int> selectedLevels)
         {
@@ -52,8 +52,8 @@ namespace SFA.DAS.FAT.Web.UnitTests.Models.CoursesViewModelTests
             var buildLevelsLink = "";
 
             // Act
-            var model = CoursesViewModelFactory.BuildModel(new List<Guid>(), null, selectedLevels, "name");
-            
+            var model = CoursesViewModelFactory.BuildModel(new List<Guid>(), null, selectedLevels, "relevance");
+
             foreach (var selectedLevel in selectedLevels)
             {
                 var separator = string.IsNullOrEmpty(buildLevelsLink) ? "?" : "&";
@@ -61,17 +61,23 @@ namespace SFA.DAS.FAT.Web.UnitTests.Models.CoursesViewModelTests
             }
 
             // Assert​
-            model.OrderByName.Should().Be($"{buildLevelsLink}&orderby=name");
+            model.OrderByRelevance.Should().Be($"{buildLevelsLink}&orderby=relevance");
         }
+
         [Test, AutoData]
-        public void And_Then_Adds_Sectors_And_Levels_To_Query_String(List<int> selectedLevels, List<Guid> selectedRoutes)
+        public void Then_Adds_Sectors_And_Levels_To_Query_String(List<int> selectedLevels, List<Guid> selectedRoutes)
         {
             // Arrange 
             var buildLevelsLink = "";
             var buildSectorLink = "";
-           
+
             // Act
-            var model = CoursesViewModelFactory.BuildModel(selectedRoutes, null, selectedLevels, "name");
+            var model = CoursesViewModelFactory.BuildModel(selectedRoutes, null, selectedLevels, "relevance");
+            foreach (var selectedLevel in selectedLevels)
+            {
+                var separator = string.IsNullOrEmpty(buildLevelsLink) ? "?" : "&";
+                buildLevelsLink += model.SelectedLevels != null && model.SelectedLevels.Any() ? $"{separator}levels=" + string.Join("levels=", selectedLevel) : "";
+            }
 
             foreach (var selectedRoute in selectedRoutes)
             {
@@ -79,14 +85,8 @@ namespace SFA.DAS.FAT.Web.UnitTests.Models.CoursesViewModelTests
                 buildSectorLink += model.SelectedSectors != null && model.SelectedSectors.Any() ? $"{separator}sectors=" + string.Join("&sectors=", selectedRoute) : "";
             }
 
-            foreach (var selectedLevel in selectedLevels)
-            {
-                var separator = string.IsNullOrEmpty(buildLevelsLink) ? "?" : "&";
-                buildLevelsLink += model.SelectedLevels != null && model.SelectedLevels.Any() ? $"{separator}levels=" + string.Join("&levels=", selectedLevel) : "";
-            }
-
             // Assert
-            model.OrderByName.Should().Be($"{buildSectorLink}{buildLevelsLink}&orderby=name");
+            model.OrderByRelevance.Should().Be($"{buildSectorLink}{buildLevelsLink}&orderby=relevance");
         }
     }
 }
